general: split elektra folder into core and kdb
also extract base and pluginload object libraries
