+++ src/include/CMakeLists.txt
@@ -19,7 +19,6 @@ check_include_file (ctype.h HAVE_CTYPE_H)
 check_include_file (errno.h HAVE_ERRNO_H)
 check_include_file (features.h HAVE_FEATURES_H)
 check_include_file (locale.h HAVE_LOCALE_H)
-check_include_file (stdio.h HAVE_STDIO_H)
 check_include_file (stdlib.h HAVE_STDLIB_H)
 check_include_file (string.h HAVE_STRING_H)
 check_include_file (time.h HAVE_TIME_H)
