+++ src/libs/base/nolog.c
