+++ src/libs/core/cow.c
