+++ src/libs/core/errors.c
