+++ src/libs/core/kdbenum.c
