+++ src/libs/core/opmphm.c
