+++ src/libs/core/proposal.c
