+++ src/libs/core/rand.c
