+++ src/libs/kdb/changetracking.c
