+++ src/libs/kdb/contracts.c
