+++ src/libs/kdb/diff.c
