+++ src/libs/kdb/hooks.c
