+++ src/libs/pluginload/doc.c
