+++ src/libs/pluginload/static.c
