+++ tests/ctest/test_opmphm_predictor.c
@@ -6,7 +6,7 @@
  * @copyright BSD License (see doc/LICENSE.md or https://www.libelektra.org)
  */
 
-#include "../../src/libs/elektra/opmphmpredictor.c"
+#include "../../src/libs/core/opmphmpredictor.c"
 
 #include <tests_internal.h>
